+++ lib/helpers/payment_tags.dart
@@ -7,5 +7,14 @@ const kCardPaymentOptions = <String>{
   // Credit cards
   'credit_cards',
   'american_express', 'diners_club', 'discover_card', 'jcb',
-  'mastercard', 'unionpay', 'visa',
+  'mastercard', 'unionpay', 'visa', 'mir', 'belkart', 'pro100',
+  // Contactless
+  'contactless',
+  'mastercard_contactless', 'paypass', 'visa_contactless',
+  'girocard_contactless', 'quickpass', 'QUICpay',
+  'apple_pay', 'google_pay'
+  // Digital wallets
+  'troika', 'wechat', 'alipay', 'blik', 'gcash', 'touchngo',
+  'huawei_pay', 'interac', 'line_pay', 'mipay', 'samsung_pay',
+  'satispay', 'swish', 'twint',
 };