+++ lib/screens/editor/entrance.dart
@@ -92,7 +92,7 @@ class _EntranceEditorPaneState extends ConsumerState<EntranceEditorPane> {
         }
       }
     }
-    return ['1', '3', '4', '5'];
+    return ['1', '2', '3', '4'];
   }
 
   @override