+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 _Unreleased_
 
+* Micromapping mode now has coloured dots instead of tiles and numbers.
 * Added app version to the settings screen.
 * Fixed location scopes in the app.
 * Made hint labels lighter.
+* Button in hours editor to use the most common value from around.
 
 ## 0.1.7